customer-review/app/requirements.txt
httpx==0.23.0
fastapi==0.104.1
pydantic==1.10.2
python-dotenv==0.21.0
uvicorn==0.24.0
requests==2.31.0
pytest==7.2.0
deepdiff==6.2.1
google-auth==2.14.1
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
pandas==1.5.1
cryptography==38.0.3
scikit-learn==1.1.3
plotly-express==0.4.1
torch==2.1.0
transformers==4.35.0